requirements.txt
nautilus_trader
pandas
ib_async
loguru
python-logging-loki
asyncio
pyyaml
google-generativeai
python-dotenv
numpy
pytest
pyarrow
